With nosync flag

+++ run.sh
@@ -7,8 +7,13 @@ LOCAL_VOLUME_PATH=${LOCAL_VOLUME_PATH:-/data/local}
 REMOTE_VOLUME_PATH=${REMOTE_VOLUME_PATH:-/data/remote}
 SNAPSHOT_INTERVAL=${SNAPSHOT_INTERVAL:-120}
 
+DISABLE_SYNC_FLAG="$LOCAL_VOLUME_PATH/nosync"
+
+INITIAL_SYNC_DELAY=5
+
 force_sync() {
   echo "Replication now signal"
+  rm -f "$DISABLE_SYNC_FLAG"
   killall -SIGTERM sleep || echo "No replication task available"
 }
 
@@ -28,14 +33,20 @@ abort(){
 echo "Initial sync from $REMOTE_VOLUME_PATH"
 
 crash="$LOCAL_VOLUME_PATH-crash-`date +"%Y-%m-%d-%H-%M-%S"`"
-# first, make a backup
-cp -al "$LOCAL_VOLUME_PATH" "$crash"
 
-rsync -av --delete "$REMOTE_VOLUME_PATH"/ "$LOCAL_VOLUME_PATH"/
+
+if [ ! -f "$DISABLE_SYNC_FLAG" ] ; then
+  cp -al "$LOCAL_VOLUME_PATH" "$crash"
+  rsync -av --delete "$REMOTE_VOLUME_PATH"/ "$LOCAL_VOLUME_PATH"/
+else
+  echo "Skipping initial sync"
+  INITIAL_SYNC_DELAY=infinity
+fi
+
 
 
 background_sync(){
-  sleep 5 # wait for prometheus to be ready (todo, better)
+  sleep $INITIAL_SYNC_DELAY || true # wait for prometheus to be ready (todo, better)
 
   while true; do
   echo "Now running snapshot"
